+++ src/main/java/nl/esciencecenter/xenon/cli/listfiles/ListFilesOutput.java
@@ -1,50 +1,27 @@
 package nl.esciencecenter.xenon.cli.listfiles;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
-
 import nl.esciencecenter.xenon.filesystems.Path;
 import nl.esciencecenter.xenon.filesystems.PathAttributes;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * Listing of files and directories of a location
  */
 public class ListFilesOutput {
-    private List<String> objects = new ArrayList<>();
-    private List<String> files = new ArrayList<>();
-    private List<String> directories = new ArrayList<>();
+    private final List<String> files;
 
-    ListFilesOutput(Path start, Iterable<PathAttributes> items, Boolean hidden) {
-        // keep track of all hidden dirs,
-        // if path starts with one of the hidden dirs than path is also hidden
-        // only works when parent dir is given before child
-        Set<Path> hiddenDirs = new HashSet<>();
-        for (PathAttributes item: items) {
-            String path = start.relativize(item.getPath()).getRelativePath();
-            if (!hidden && item.isDirectory() && item.isHidden()) {
-                hiddenDirs.add(item.getPath());
-            }
-            // the current item is hidden or one of its parents is hidden
-            boolean parentIsHidden = hiddenDirs.stream().anyMatch(d -> item.getPath().startsWith(d));
-            boolean itemIsHidden = item.isHidden() || parentIsHidden;
-            if (!itemIsHidden || hidden) {
-                objects.add(path);
-                if (item.isDirectory()) {
-                    directories.add(path);
-                } else if (item.isRegular()) {
-                    files.add(path);
-                }
-            }
-        }
+    public ListFilesOutput(Path start, Stream<PathAttributes> stream) {
+        files = stream.map(item -> start.relativize(item.getPath()).getRelativePath()).collect(Collectors.toList());
     }
 
     @Override
     public String toString() {
         String sep = System.getProperty("line.separator");
-        return String.join(sep, objects);
+        return String.join(sep, files);
     }
 
     @Override
@@ -56,17 +33,11 @@ public class ListFilesOutput {
             return false;
         }
         ListFilesOutput that = (ListFilesOutput) o;
-        return Objects.equals(objects, that.objects) &&
-                Objects.equals(files, that.files) &&
-                Objects.equals(directories, that.directories);
+        return Objects.equals(files, that.files);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(objects, files, directories);
-    }
-
-    public List<String> getObjects() {
-        return objects;
+        return Objects.hash(files);
     }
 }